choosing testing in enviroments

+++ scripts/test-cf-openvidu-all-3.sh
@@ -60,26 +60,28 @@ else
   exit 0
 fi  
 
-aws cloudformation create-stack \
-  --stack-name Openvidu-selfsigned-${DOMAIN_NAME} \
-  --template-url ${CF_FILE} \
-  --parameters file:///${TEMPJSON} \
-  --disable-rollback
+if [ "$MODE" == "dev" ]; then
+  aws cloudformation create-stack \
+    --stack-name Openvidu-selfsigned-${DOMAIN_NAME} \
+    --template-url ${CF_FILE} \
+    --parameters file:///${TEMPJSON} \
+    --disable-rollback
 
-aws cloudformation wait stack-create-complete --stack-name Openvidu-selfsigned-${DOMAIN_NAME}
+  aws cloudformation wait stack-create-complete --stack-name Openvidu-selfsigned-${DOMAIN_NAME}
 
-echo "Extracting service URL..."
-URL=$(aws cloudformation describe-stacks --stack-name Openvidu-selfsigned-${DOMAIN_NAME} | jq -r '.Stacks[0] | .Outputs[] | select(.OutputKey | contains("WebsiteURL")) | .OutputValue')
+  echo "Extracting service URL..."
+  URL=$(aws cloudformation describe-stacks --stack-name Openvidu-selfsigned-${DOMAIN_NAME} | jq -r '.Stacks[0] | .Outputs[] | select(.OutputKey | contains("WebsiteURL")) | .OutputValue')
 
-sleep 10
-RES=$(curl --insecure --location -u OPENVIDUAPP:MY_SECRET --output /dev/null --silent --write-out "%{http_code}\\n" ${URL} | grep "200")
+  sleep 10
+  RES=$(curl --insecure --location -u OPENVIDUAPP:MY_SECRET --output /dev/null --silent --write-out "%{http_code}\\n" ${URL} | grep "200")
 
-# Cleaning up
-aws cloudformation delete-stack --stack-name Openvidu-selfsigned-${DOMAIN_NAME}
+  # Cleaning up
+  aws cloudformation delete-stack --stack-name Openvidu-selfsigned-${DOMAIN_NAME}
 
-if [ "$RES" != "200" ]; then
-  echo "deployment failed"
-  exit 1
+  if [ "$RES" != "200" ]; then
+    echo "deployment failed"
+    exit 1
+  fi
 fi
 
 #############################
@@ -155,27 +157,28 @@ else
   exit 0
 fi
 
-aws cloudformation create-stack \
-  --stack-name Openvidu-owncert-${DOMAIN_NAME} \
-  --template-url ${CF_FILE} \
-  --parameters file:///$TEMPJSON \
-  --disable-rollback
+if [ "$MODE" = "dev" ]; then
+  aws cloudformation create-stack \
+    --stack-name Openvidu-owncert-${DOMAIN_NAME} \
+    --template-url ${CF_FILE} \
+    --parameters file:///$TEMPJSON \
+    --disable-rollback
 
-aws cloudformation wait stack-create-complete --stack-name Openvidu-owncert-${DOMAIN_NAME}
+  aws cloudformation wait stack-create-complete --stack-name Openvidu-owncert-${DOMAIN_NAME}
 
-echo "Extracting service URL..."
-URL=$(aws cloudformation describe-stacks --stack-name Openvidu-owncert-${DOMAIN_NAME} | jq -r '.Stacks[0] | .Outputs[] | select(.OutputKey | contains("WebsiteURLLE")) | .OutputValue')
+  echo "Extracting service URL..."
+  URL=$(aws cloudformation describe-stacks --stack-name Openvidu-owncert-${DOMAIN_NAME} | jq -r '.Stacks[0] | .Outputs[] | select(.OutputKey | contains("WebsiteURLLE")) | .OutputValue')
 
-sleep 10
-RES=$(curl --insecure --location -u OPENVIDUAPP:MY_SECRET --output /dev/null --silent --write-out "%{http_code}\\n" ${URL} | grep "200")
+  sleep 10
+  RES=$(curl --insecure --location -u OPENVIDUAPP:MY_SECRET --output /dev/null --silent --write-out "%{http_code}\\n" ${URL} | grep "200")
 
-# Cleaning up
-aws cloudformation delete-stack --stack-name Openvidu-owncert-${DOMAIN_NAME}
+  # Cleaning up
+  aws cloudformation delete-stack --stack-name Openvidu-owncert-${DOMAIN_NAME}
 
-sleep 60
+  sleep 60
 
-ALLOCATION_ID=$(aws ec2 describe-addresses --public-ips ${IP} | jq -r ' .Addresses[0] | .AllocationId')
-aws ec2 release-address --allocation-id ${ALLOCATION_ID} 
+  ALLOCATION_ID=$(aws ec2 describe-addresses --public-ips ${IP} | jq -r ' .Addresses[0] | .AllocationId')
+  aws ec2 release-address --allocation-id ${ALLOCATION_ID} 
 
 cat >$TEMPFILE<<EOF
 {
@@ -198,12 +201,13 @@ cat >$TEMPFILE<<EOF
 }
 EOF
 
-aws route53 change-resource-record-sets --hosted-zone-id ZVWKFNM0CR0BK \
-  --change-batch file:///$TEMPFILE
+  aws route53 change-resource-record-sets --hosted-zone-id ZVWKFNM0CR0BK \
+    --change-batch file:///$TEMPFILE
 
-if [ "$RES" != "200" ]; then
-  echo "deployment failed"
-  exit 1
+  if [ "$RES" != "200" ]; then
+    echo "deployment failed"
+    exit 1
+  fi
 fi
 
 #############################
@@ -270,27 +274,28 @@ else
   exit 0
 fi
 
-aws cloudformation create-stack \
-  --stack-name Openvidu-letsencrypt-${DOMAIN_NAME} \
-  --template-url ${CF_FILE} \
-  --parameters file:///$TEMPJSON \
-  --disable-rollback
+if [ "$MODE" == "dev" ]; then
+  aws cloudformation create-stack \
+    --stack-name Openvidu-letsencrypt-${DOMAIN_NAME} \
+    --template-url ${CF_FILE} \
+    --parameters file:///$TEMPJSON \
+    --disable-rollback
 
-aws cloudformation wait stack-create-complete --stack-name Openvidu-letsencrypt-${DOMAIN_NAME}
+  aws cloudformation wait stack-create-complete --stack-name Openvidu-letsencrypt-${DOMAIN_NAME}
 
-echo "Extracting service URL..."
-URL=$(aws cloudformation describe-stacks --stack-name Openvidu-letsencrypt-${DOMAIN_NAME} | jq -r '.Stacks[0] | .Outputs[] | select(.OutputKey | contains("WebsiteURLLE")) | .OutputValue')
+  echo "Extracting service URL..."
+  URL=$(aws cloudformation describe-stacks --stack-name Openvidu-letsencrypt-${DOMAIN_NAME} | jq -r '.Stacks[0] | .Outputs[] | select(.OutputKey | contains("WebsiteURLLE")) | .OutputValue')
 
-sleep 10
-RES=$(curl --location -u OPENVIDUAPP:MY_SECRET --output /dev/null --silent --write-out "%{http_code}\\n" ${URL} | grep "200")
+  sleep 10
+  RES=$(curl --location -u OPENVIDUAPP:MY_SECRET --output /dev/null --silent --write-out "%{http_code}\\n" ${URL} | grep "200")
 
-# Cleaning up
-aws cloudformation delete-stack --stack-name Openvidu-letsencrypt-${DOMAIN_NAME}
+  # Cleaning up
+  aws cloudformation delete-stack --stack-name Openvidu-letsencrypt-${DOMAIN_NAME}
 
-sleep 60
+  sleep 60
 
-ALLOCATION_ID=$(aws ec2 describe-addresses --public-ips ${IP} | jq -r '.Addresses[0] | .AllocationId')
-aws ec2 release-address --allocation-id ${ALLOCATION_ID} 
+  ALLOCATION_ID=$(aws ec2 describe-addresses --public-ips ${IP} | jq -r '.Addresses[0] | .AllocationId')
+  aws ec2 release-address --allocation-id ${ALLOCATION_ID} 
 
 cat >$TEMPFILE<<EOF
 {
@@ -313,12 +318,13 @@ cat >$TEMPFILE<<EOF
 }
 EOF
 
-aws route53 change-resource-record-sets --hosted-zone-id ZVWKFNM0CR0BK \
-  --change-batch file:///$TEMPFILE
+  aws route53 change-resource-record-sets --hosted-zone-id ZVWKFNM0CR0BK \
+    --change-batch file:///$TEMPFILE
 
-if [ "$RES" != "200" ]; then
-  echo "deployment failed"
-  exit 1
+  if [ "$RES" != "200" ]; then
+    echo "deployment failed"
+    exit 1
+  fi
 fi
 
 # Cleaning
